+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `@Feature` and `@Action macros.
 - HeroList pagination.
+- HeroList unit tests.
 
 ### Fixed